+++ changelog.md
@@ -4,6 +4,8 @@ dev
 ------------------
 - fix bug: multiple UnitRegistries conflicted
 - fix bug: Comparison.implies(Predicate) should be False
+- merge `comparable` module into `terms` module
+- fix bug: Comparison >= Predicate should be False
 
 0.2.0 (2021-03-02)
 ------------------